exec: add the hashed aggregator
hashedAggregator is an operator that performs an aggregation based on specified
grouping columns. This operator performs a blocking hash table build at the
beginning and then the grouped buckets are then passed as input to an
orderedAggregator. The output row ordering of this aggregator is arbitrary.

Release note: None

+++ pkg/sql/exec/sort.go
@@ -143,7 +143,7 @@ func (p *sortOp) Next() ColBatch {
 		p.state = sortEmitting
 		fallthrough
 	case sortEmitting:
-		newEmitted := p.emitted + ColBatchSize
+		newEmitted := p.emitted + uint64(ColBatchSize)
 		if newEmitted > p.spooledTuples {
 			newEmitted = p.spooledTuples
 		}
